requirements.txt
astral==3.2
customtkinter==5.2.2
ffmpeg-python==0.2.0
opencv-python==4.9.0.80
pillow==11.0.0
plyer>=2.1.0
pytz==2024.1
pystray==0.19.5
requests==2.32.3
tkintermapview==1.29
tzlocal>=5.0
yt-dlp==2024.1.1
